add shared animations, [refactor] move github-service and repo interface to shared/ directory
