+++ src/app/stargazers/stargazers.component.html
@@ -1,7 +1,7 @@
 <h1 class="capitalize">{{repoDetail.name}}</h1>
 <p>{{repoDetail.description}}</p>
 <div class="row">
-  <div class="stargazer col-6 col-sm-3 align-self-center" *ngFor="let stargazer of stargazers">
+  <div class="stargazer col-6 col-sm-3 align-self-center" *ngFor="let stargazer of stargazers" [@slideInUp]="'in'">
     <img class="img-fluid" [src]="stargazer" />
   </div>
   <div class="col-12">
